+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - **Breaking**: `Array` `retrieve_` methods now return `Vec<u8>`/`Vec<T>` instead of `Box<[u8]>`/`Box<[T]>`
    - This avoids potential internal reallocations
+ - **Breaking**: Remove `StorePrefixError::new`, deprecated since `v0.7.3`
 
 ## [0.10.0] - 2024-01-17